Refactored API structure to enhance readability and facilitating the integration of future device models
