+++ custom_components/ambeo_soundbar/__init__.py
@@ -6,8 +6,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.exceptions import ConfigEntryNotReady
 from homeassistant.helpers import device_registry as dr
 
-from .const import DOMAIN
-from .api import AmbeoApi
+from .const import DEFAULT_PORT, DOMAIN, MANUFACTURER, MAX_SOUNDBAR
+from .api.factory import AmbeoAPIFactory
 
 
 _LOGGER = logging.getLogger(__name__)
@@ -22,31 +22,24 @@ class AmbeoDevice:
         self.version = version
         self.host = host
         self.port = port
-        
+
     @property
     def serial(self):
         return self._serial
-    
-    @property
-    def max_compat(self):
-        return self.model == "AMBEO Soundbar Max"
 
 
-async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry):    
+async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry):
     _LOGGER.debug("Démarrage de la configuration du composant ambeo")
     config = entry.data
     host = config.get("host")
-    port = 80
-    ambeo_api = AmbeoApi(host, port, aiohttp.ClientSession(), hass)
-    serial= await ambeo_api.get_serial()
-    if serial is None:
-        raise ConfigEntryNotReady(f"Can't connect to host : {host}") 
-    model= await ambeo_api.get_model()
-    name= await ambeo_api.get_name()
-    version= await ambeo_api.get_version()
-    manufacturer= "Sennheiser"
-    device = AmbeoDevice(serial, name, manufacturer, model, version, host, port)
-    ambeo_api._ambeo_max_compat = device.max_compat
+    ambeo_api = await AmbeoAPIFactory.create_api(
+        host, DEFAULT_PORT, aiohttp.ClientSession(), hass)
+    serial = await ambeo_api.get_serial()
+    model = await ambeo_api.get_model()
+    name = await ambeo_api.get_name()
+    version = await ambeo_api.get_version()
+    device = AmbeoDevice(serial, name, MANUFACTURER,
+                         model, version, host, DEFAULT_PORT)
     hass.data.setdefault(DOMAIN, {})
     _LOGGER.debug("Data initialized")
     hass.data[DOMAIN][entry.entry_id] = {
@@ -55,12 +48,12 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry):
     }
     device_registry = dr.async_get(hass)
     device_registry.async_get_or_create(
-         config_entry_id=entry.entry_id,
-         identifiers={(DOMAIN, serial)},
-         name=name,
-         manufacturer=manufacturer,
-         model=model,
-         sw_version=version)
+        config_entry_id=entry.entry_id,
+        identifiers={(DOMAIN, serial)},
+        name=name,
+        manufacturer=MANUFACTURER,
+        model=model,
+        sw_version=version)
 
     hass.async_create_task(
         hass.config_entries.async_forward_entry_setups(
