+++ custom_components/ambeo_soundbar/const.py
@@ -1,13 +1,30 @@
 DOMAIN = "ambeo_soundbar"
 VERSION = "1.0.0"
-DEFAULT_BRIGHTNESS = 50
+MANUFACTURER = "Sennheiser"
+DEFAULT_PORT = 80
+
 BRIGHTNESS_SCALE = (0, 100)
-BRIGHTNESS_SCALE_AMBEO_MAX_LOGO = (1, 118)
+DEFAULT_BRIGHTNESS = 50
+
+# MAX
+MAX_SOUNDBAR = "AMBEO Soundbar Max"
 BRIGHTNESS_SCALE_AMBEO_MAX_DISPLAY = (1, 126)
 EXCLUDE_SOURCES_MAX = ["aes"]
-AMBEO_PLUS_VOLUME_STEP = 0.01
 AMBEO_MAX_VOLUME_STEP = 0.02
-
 DEFAULT_BRIGHTNESS_AMBEO_MAX = 128
+BRIGHTNESS_SCALE_AMBEO_MAX_LOGO = (1, 118)
+
+# PLUS
+PLUS_SOUNDBAR = "AMBEO Soundbar Plus"
+AMBEO_PLUS_VOLUME_STEP = 0.01
 
 
+class Capability:
+    AMBEO_LOGO = "AmbeoLogo"
+    LED_BAR = "LEDBar"
+    CODEC_LED = "CodecLED"
+    VOICE_ENHANCEMENT = "VoiceEnhancementMode"
+    BLUETOOTH_PAIRING = "AmbeoBluetoothPairing"
+    STANDBY = "standby"
+    MAX_LOGO = "AmbeoMaxLogo"
+    MAX_DISPLAY = "AmbeoMaxDisplay"
