+++ custom_components/ambeo_soundbar/manifest.json
@@ -9,5 +9,5 @@
   "iot_class": "local_polling",
   "issue_tracker": "https://github.com/faizpuru/ha-ambeo_soundbar",
   "requirements": [],
-  "version": "1.0.0"
+  "version": "1.0.2"
 }
